Refactor: Dev merged in and conflicts resolution
