+++ src/app/page.tsx
@@ -3,6 +3,7 @@ import HowItWorks from "~/components/layouts/homepage/HowItWorks";
 import PerfectFit from "~/components/layouts/homepage/PerfectFit";
 import Testimonials from "~/components/layouts/homepage/Testimonials";
 import UserSection from "~/components/layouts/homepage/UserSection";
+import TopProductLists from "~/components/TopProductLists/TopProductLists";
 
 export default function Home() {
   return (
@@ -12,6 +13,7 @@ export default function Home() {
       <HowItWorks />
       <Testimonials />
       <PerfectFit />
+      <TopProductLists />
     </>
   );
 }
